Change order of import based on PEP8, isort, & Patrick's recommendations
